+++ mex/editor/client/src/app/fast-track-resource/vocabulary-search.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from "@angular/core/testing";
+
+import { VocabularySearch } from "./vocabulary-search.service";
+
+describe("VocabularySearchService", () => {
+  let service: VocabularySearch;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(VocabularySearch);
+  });
+
+  it("should be created", () => {
+    expect(service).toBeTruthy();
+  });
+});