+++ mex/editor/client/src/app/shared/backend-proxy.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from "@angular/core/testing";
+
+import { BackendProxy } from "./backend-proxy.service";
+
+describe("BackendSearchService", () => {
+  let service: BackendProxy;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(BackendProxy);
+  });
+
+  it("should be created", () => {
+    expect(service).toBeTruthy();
+  });
+});