+++ mex/editor/client/src/app/shared/models/index.ts
@@ -1,0 +1,9 @@
+import type { PreviewOrganizationalUnit } from "./organizational-unit";
+
+export type { PreviewOrganizationalUnit } from "./organizational-unit";
+export type { Text, TextLanguage, RestrictedTextLanguage } from "./text";
+
+/**
+ * UnionType for all preview items.
+ */
+export type PreviewItem = PreviewOrganizationalUnit;