+++ mex/editor/client/src/app/start-page/start-page.ts
@@ -1,0 +1,14 @@
+import { Component } from "@angular/core";
+import { MatButton } from "@angular/material/button";
+import { RouterLink } from "@angular/router";
+
+@Component({
+  selector: "mex-start-page",
+  imports: [RouterLink, MatButton],
+  templateUrl: "./start-page.html",
+  styleUrl: "./start-page.scss",
+})
+/**
+ * StartPage for the app.
+ */
+export class StartPage {}